Crate drgrep

drgrep

A Rust implementation of the grep software with more support and features for args, workspace scanning and CLI.

Features

Recursive research

Command Lines Parser

Regex Utilities

CLI coloration

Examples

use drgrep::{args::parser::ArgParser, search_sensitive_case};

fn main() {
    let args = ArgParser::new();
    println!("Results: {:?}", args);
    let search_key = "duct";
    let content = "\
Rust:
sécurité, rapidité, productivité.
Obtenez les trois en même temps.
Duck tape.";
    assert_eq!(
    vec!["sécurité, rapidité, productivité."],
    search_sensitive_case(search_key, content)
);
}

Re-exports

pub use args::parser::ArgParser;

pub use color::config::Color;

pub use color::printer::print_colored;

pub use color::printer::print_partial_colored;

pub use color::printer::print_styled;

pub use regex::pattern::find;

pub use regex::pattern::find_all;

pub use regex::pattern::is_match;

pub use regex::pattern::replace_all;

Modules

args

ArgParser Module

color

glob

Enhanced Glob Pattern Matcher

regex

temp_dir

utilities 🔒

Macros

print_colored

Macro for printing colored text

print_partial_colored

Macro for printing multiple colored text segments

print_styled

Macro for printing styled and colored text

Structs

Config

The config struct

SearchResult

Statics

DEFAULT_MESSAGE

Functions

run

search_insensitive_case

search_sensitive_case

search_with_regex

search_word_insensitive_case

search_word_sensitive_case