Crate drgrep

drgrep

A Rust implementation of the grep software with more support and features for args, workspace scanning and CLI.

Features

Recursive research

Command Lines Parser

Regex Utilities

CLI coloration

Examples

use drgrep::{args::parser::ArgParser, search_sensitive_case};

fn main() {
    let args = ArgParser::new();
    println!("Results: {:?}", args);
    let search_key = "duct";
    let content = "\
Rust:
sécurité, rapidité, productivité.
Obtenez les trois en même temps.
Duck tape.";
    assert_eq!(
    vec!["sécurité, rapidité, productivité."],
    search_sensitive_case(search_key, content)
);
}

Re-exports

pub use color::config::Color;

pub use color::printer::print_colored;

pub use color::printer::print_partial_colored;

pub use color::printer::print_styled;

Modules

args

color

regex

Structs

Config

The config struct

LinesInfo

SearchResult

Functions

can_read_to_utf8 🔒

run

search_insensitive_case

search_sensitive_case

search_word_insensitive_case

search_word_sensitive_case